dont set log_facility during config validation
the validation function is not called when a default value is used.
with the old implementation set_log_facility() is never called.
if the default value for the config and the hard coded default value
for the DEFAULT_LOG_FACILITY differ first the DEFAULT_LOG_FACILITY
will be used and if a module sets the log_facility from config file
the default log facility from the config will be used.

now the validation only checks if the value is a valid one and the
base file always sets the default log facility that is obtained
from the config.